requirements.txt
bottle==0.12.16
PyYAML==3.13
